Update help message, error message, commands
The help message now mentions that raw unbuffered writes are used in dd command. The error message usage statement dynamically updates based on the name of the command, this was required to prevent confusion if the error was not updated following a change in the script's name. hdiutil now has the ISO file immediately following the convert verb, as suggested by hdiutil man page.dd command has block size of 4m instead of 1m and is followed by sync command if it succeeds.